added ConfirmationDialog before deleting

+++ app/src/main/java/com/example/tvshowlist/presentation/ui/HomeScreen.kt
@@ -1,9 +1,6 @@
 package com.example.tvshowlist.presentation.ui
 
-import androidx.compose.animation.animateColorAsState
-import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -18,12 +15,9 @@ import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Clear
-import androidx.compose.material.icons.filled.Delete
 import androidx.compose.material.icons.filled.Search
 import androidx.compose.material.icons.filled.Settings
 import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.DismissDirection
-import androidx.compose.material3.DismissValue
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
@@ -31,20 +25,18 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarDuration
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
-import androidx.compose.material3.SwipeToDismiss
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextField
 import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.rememberDismissState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.scale
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
@@ -57,6 +49,7 @@ import androidx.compose.ui.unit.sp
 import com.example.tvshowlist.R
 import com.example.tvshowlist.domain.model.TvShow
 import com.example.tvshowlist.presentation.MainViewModel
+import com.example.tvshowlist.presentation.ui.items.ConfirmationDialog
 import com.example.tvshowlist.presentation.ui.items.ItemTvShow
 import com.example.tvshowlist.presentation.ui.items.SwipeToDeleteContainer
 import com.example.tvshowlist.utils.ApplicationOnlineChecker
@@ -195,10 +188,28 @@ fun HomeScreen(
                                         navigateTo = navigateTo
                                     )
                                 } else {
+                                    var showDeletionDialog by remember { mutableStateOf(false) }
+                                    var isRecentShowRemoved by remember { mutableStateOf(false) }
+
+                                    ConfirmationDialog(
+                                        showDialog = showDeletionDialog,
+                                        title = context.getString(R.string.recent_tv_show_deletion_title),
+                                        message = context.getString(R.string.delete_recent_tv_show_confirmation),
+                                        confirmButtonText = context.getString(R.string.delete),
+                                        onConfirm = {
+                                            viewModel.deleteRecentTvShow(tvShow)
+                                            isRecentShowRemoved = true
+                                        },
+                                        onDismiss = {
+                                            showDeletionDialog = false
+                                            isRecentShowRemoved = false
+                                        }
+                                    )
                                     SwipeToDeleteContainer(
                                         item = tvShow,
+                                        isRecentShowRemoved = isRecentShowRemoved,
                                         onDelete = {
-                                            viewModel.deleteRecentTvShow(tvShow)
+                                            showDeletionDialog = true
                                         },
                                         animationDuration = 500
                                     ) {
